完善了修改信息,修改广告,管理商品模块的逻辑跳转等,差与界面交互部分,修改了商品信息中价格从long型换到double型

+++ src/main/java/com/project_management/shoppingweb/domain/ProductInformation.java
@@ -84,7 +84,7 @@ public class ProductInformation {
         return seller;
     }
 
-    public ProductInformation(long productId, long seller, String productName, long productNumber, long productPrice, String JSONStr, String productPicture, String productNote) {
+    public ProductInformation(long productId, long seller, String productName, long productNumber, double productPrice, String JSONStr, String productPicture, String productNote) {
         this.productId = productId;
         this.seller = seller;
         this.productName = productName;
